Use WhisperKit (#36)

+++ Sources/AppService/Service.swift
@@ -60,7 +60,33 @@ public final class Service {
 
   }
 
-  public func importItem(title: String, audioFileURL: URL, segments: [AbstractSegment]) async throws {
+  public func updateTranscribe(for item: ItemEntity) async throws {
+
+    let result = try await WhisperKitWrapper.run(url: item.audioFileAbsoluteURL)
+    try await self.importItem(
+      title: item.title,
+      audioFileURL: result.audioFileURL,
+      segments: result.segments
+    )
+
+  }
+
+  public func transcribe(title: String, audioFileURL: URL) async throws {
+
+    let result = try await WhisperKitWrapper.run(url: audioFileURL)
+    try await self.importItem(
+      title: title,
+      audioFileURL: audioFileURL,
+      segments: result.segments
+    )
+
+  }
+
+  public func importItem(
+    title: String,
+    audioFileURL: URL,
+    segments: [AbstractSegment]
+  ) async throws {
 
     let storedSubtitle = StoredSubtitle(items: segments)
 
@@ -90,6 +116,8 @@ public final class Service {
       }
 
       func overwrite(file: URL, to url: URL) throws {
+
+        guard file != url else { return }
 
         if fileManager.fileExists(atPath: url.path(percentEncoded: false)) {
           try fileManager.removeItem(at: url)
@@ -121,7 +149,16 @@ public final class Service {
           audioFileDestinationPath.relative(basedOn: .init(url: URL.documentsDirectory)).rawValue
         try new.setSegmentData(storedSubtitle)
 
+        new.pinItems = []
+
         modelContext.insert(new)
+
+        let pins = try modelContext.fetch(.init(predicate: #Predicate<PinEntity> { [identifier = new.identifier] in $0.item?.identifier == identifier }))
+
+        for pin in pins {
+          pin.item = nil
+          modelContext.delete(pin)
+        }
 
       }
     }
